requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.0.0
    # via mediapipe
annotated-types==0.6.0
    # via pydantic
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
attrs==23.1.0
    # via mediapipe
certifi==2023.7.22
    # via requests
cffi==1.16.0
    # via sounddevice
charset-normalizer==3.3.1
    # via requests
contourpy==1.1.1
    # via matplotlib
cramjam==2.7.0
    # via fastparquet
cycler==0.12.1
    # via matplotlib
fastparquet==2023.10.1
    # via -r requirements.in
filelock==3.13.0
    # via
    #   torch
    #   triton
flatbuffers==23.5.26
    # via mediapipe
fonttools==4.43.1
    # via matplotlib
fsspec==2023.10.0
    # via
    #   fastparquet
    #   torch
hydra-core==1.3.2
    # via -r requirements.in
idna==3.4
    # via requests
jinja2==3.1.2
    # via torch
kiwisolver==1.4.5
    # via matplotlib
markupsafe==2.1.3
    # via jinja2
matplotlib==3.8.0
    # via
    #   -r requirements.in
    #   mediapipe
mediapipe==0.10.7
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
networkx==3.2.1
    # via torch
numpy==1.26.1
    # via
    #   -r requirements.in
    #   contourpy
    #   fastparquet
    #   matplotlib
    #   mediapipe
    #   opencv-contrib-python
    #   opencv-python
    #   pandas
    #   scipy
    #   torchvision
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==8.9.2.26
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.18.1
    # via torch
nvidia-nvjitlink-cu12==12.3.52
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
omegaconf==2.3.0
    # via
    #   -r requirements.in
    #   hydra-core
opencv-contrib-python==4.8.1.78
    # via mediapipe
opencv-python==4.8.1.78
    # via -r requirements.in
packaging==23.2
    # via
    #   fastparquet
    #   hydra-core
    #   matplotlib
pandas==2.1.1
    # via
    #   -r requirements.in
    #   fastparquet
pillow==10.1.0
    # via
    #   matplotlib
    #   torchvision
protobuf==3.20.3
    # via mediapipe
pycparser==2.21
    # via cffi
pydantic==2.4.2
    # via -r requirements.in
pydantic-core==2.10.1
    # via pydantic
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   -r requirements.in
    #   omegaconf
requests==2.31.0
    # via torchvision
scipy==1.11.3
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sounddevice==0.4.6
    # via mediapipe
sympy==1.12
    # via torch
torch==2.1.0
    # via
    #   -r requirements.in
    #   torchaudio
    #   torchvision
torchaudio==2.1.0
    # via -r requirements.in
torchvision==0.16.0
    # via -r requirements.in
tqdm==4.66.1
    # via -r requirements.in
triton==2.1.0
    # via torch
typing-extensions==4.8.0
    # via
    #   pydantic
    #   pydantic-core
    #   torch
tzdata==2023.3
    # via pandas
urllib3==2.0.7
    # via requests
